Inline expo::expo_client::get mod
